Backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Kafka Integration
kafka-python==2.0.2
confluent-kafka==2.3.0

# Elasticsearch
elasticsearch==8.11.0

# Object Storage (MinIO/S3)
boto3==1.34.10
minio==7.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# AI Integration (Ollama)
langchain==0.1.0
langchain-community==0.0.10

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
